--- src/module-info.java
@@ -1,5 +1,0 @@
-module cafePos {
-	exports baseSettings;
-
-	requires java.desktop;
-}
